Added Main Cards with designing
