+++ src/sass/cards.sass
@@ -1,5 +1,63 @@
-
+@import "./variables"
 /* Basic Cards */
+body
+    background-color: #eee
+.card
+    margin: 50px auto
+    height: 350px
+    border: none
+    box-shadow: 5px 7px 15px #888
+
+.cardImg
+    height: 350px
+    border-top-left-radius: 20px
+    border-bottom-left-radius: 20px
+.cardImg2
+    height: 350px
+    border-top-left-radius: 40px
+    border-bottom-left-radius: 10px
+
+.text
+    background-color: #fff
+    display: flex
+    justify-content: center
+    align-items: center
+    position: relative
+
+.ribbon 
+    width: 150px
+    height: 50px
+    color: white
+    line-height: 50px
+    padding-left: 15px
+    position: absolute
+    left: -8px
+    top: 15px
+    background: linear-gradient(to right, #6a3093, #a044ff)
+
+.ribbon:before, .ribbon:after 
+    content: ""
+    position: absolute
+
+.ribbon:before 
+    height: 0
+    width: 0
+    top: -8.5px
+    left: 0.1px
+    border-bottom: 9px solid #6a3093
+    border-left: 9px solid transparent
+
+.ribbon:after 
+    height: 0
+    width: 0
+    right: -14.5px
+    border-top: 25px solid transparent
+    border-bottom: 25px solid transparent
+    border-left: 15px solid #a044ff
 
 
+
+ 
+
+    
 /* Category Cards */
